Adjust skip/restart png offset
Moving them one pixel down increases the image comparison accuracy on my end. 

In Blegas' stream yesterday, the IGT continued to run after a skip cutscene because the `SkipCinematic` feature returned a value of `9.8` when it needed to be at least `15` in order to be detected. This change seems to increase that accuracy from `9.8` to `36`.